Remove quotes around boolean values as well as host.* variables.
This is not necessarily ideal, since we should look at the data type and make the decision based on a boolean input type.

For the host.* variables, we should either add all other variable types here, and maybe make it a regex to avoid fales positive hits, or maybe add another operator.