Check $view->compact instead of view URL param
Starting with Web version 2.8.0 the view=compact URL param has been
deprecated and removed. Before, it was used to set a certain view
to compact mode. We use the showCompact URL paramater for that now.
Note that this parameter has been there forever, so this change is
backwards compatible.

See https://github.com/Icinga/icingaweb2/pull/4164 for details.