Add `setmembers` summyary line to TemplateUsageTable
`'setmembers' => $this->getSummaryLine('object', 'o.service_set_id IS NOT NULL AND o.host_id IS NOT NULL'),` is added to TemplateUsageTable
and `objects` is set to ` $this->getSummaryLine('object', 'o.service_set_id IS NULL')`. Subsequently, ObjectTableSetMembers
is added which is used in `setmembersAction` in `TemplateController`. `ObjectTableSetMembers` uses service set name or service name
as search columns.