Unify Service/Host/UserGroup code for Objects and Forms
At least as far as possible.

For objects, moved some common parts to IcingaObjectGroup, where zone_id
is now in defaultProperties and thus those don't need to be overriden in
IcingaHostGroup and IcingaServiceGroup anymore. Similar for relations
property, which don't need to be specified in descendant classes
anymore.

For forms, created a intermediate class IcingaGroupForm (analog to the
already existing IcingaObjectGroup for objects) which extends
DirectorObjectForm, and made IcingaHostGroupForm, IcingaServiceGroupForm
and IcingaUserGroupForm extend this new class instead. Moved the
addZoneElements method to this new class, thus removing it from the
descendant classes that had it repeated.